requirements.txt
asgiref==3.8.1
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.15.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
